Remove unused page
