--- apps/prairielearn/src/ee/pages/instructorAiGenerateQuestion/instructorAiGenerateQuestion.html.ts
@@ -1,65 +1,0 @@
-import { html } from '@prairielearn/html';
-
-import { HeadContents } from '../../../components/HeadContents.html.js';
-import { Navbar } from '../../../components/Navbar.html.js';
-import { compiledScriptTag, nodeModulesAssetPath } from '../../../lib/assets.js';
-
-export function InstructorAiGenerateQuestion({ resLocals }: { resLocals: Record<string, any> }) {
-  return html`
-    <!doctype html>
-    <html lang="en">
-      <head>
-        ${HeadContents({ resLocals })} ${compiledScriptTag('question.ts')}
-        <script defer src="${nodeModulesAssetPath('mathjax/es5/startup.js')}"></script>
-      </head>
-      <body hx-ext="loading-states">
-        ${Navbar({ navPage: 'course_admin', resLocals })}
-        <main id="content" class="container mb-4">
-          <div class="mb-3">
-            <a
-              href="${resLocals.urlPrefix}/ai_generate_question_drafts"
-              class="btn btn-sm btn-primary"
-            >
-              <i class="fa fa-arrow-left" aria-hidden="true"></i>
-              Back to draft questions
-            </a>
-          </div>
-          <div class="card">
-            <div
-              class="card-header bg-primary text-white d-flex align-items-center justify-content-between"
-            >
-              Generate question with AI
-            </div>
-          </div>
-        </main>
-      </body>
-      <script>
-        function setPromptToExample() {
-          const options = document.getElementById('user-prompt-example').options;
-          const selection = options[options.selectedIndex].dataset;
-
-          document.getElementById('user-prompt-llm').value = selection.promptGeneral;
-          document.getElementById('user-prompt-llm-user-input').value = selection.promptUserInput;
-          document.getElementById('user-prompt-llm-grading').value = selection.promptGrading;
-        }
-      </script>
-    </html>
-  `.toString();
-}
-
-export function GenerationFailure({
-  urlPrefix,
-  jobSequenceId,
-}: {
-  urlPrefix: string;
-  jobSequenceId: string;
-}): string {
-  return html`
-    <div id="generation-results">
-      <h3>Generation Failed</h3>
-
-      <p>The LLM did not generate any question file.</p>
-      <a href="${urlPrefix + '/jobSequence/' + jobSequenceId}">See job logs</a>
-    </div>
-  `.toString();
-}
